# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Project PM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks (optionally filtered by project or milestone)
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: milestoneId
          in: query
          schema:
            type: integer
        - name: assigneeId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: taskId
          in: query
          description: Filter tasks attached to a parent task (reserved, unused)
          schema:
            type: integer
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Task:
      type: object
      required: [id, projectId, title, status, priority, category, createdAt]
      properties:
        id:          { type: integer }
        projectId:    { type: integer }
        milestoneId: { type: ["integer", "null"] }
        assigneeId:  { type: ["integer", "null"] }
        title:       { type: string }
        description: { type: ["string", "null"] }
        status:      { type: string, enum: [todo, in_progress, blocked, review, done] }
        priority:    { type: string, enum: [low, medium, high, critical] }
        category:    { type: string, enum: [construction, electrical, av, it, network, acoustics, furnishing, signage, general] }
        dueDate:     { type: ["string", "null"] }
        completedAt: { type: ["string", "null"] }
        createdAt:   { type: string }
        updatedAt:   { type: string }
        # Joined fields
        assigneeName:   { type: ["string", "null"] }
        projectName:     { type: ["string", "null"] }
        milestoneName:  { type: ["string", "null"] }
    TaskInput:
      type: object
      required: [projectId, title, status, priority, category]
      properties:
        projectId:    { type: integer }
        milestoneId: { type: integer }
        assigneeId:  { type: integer }
        title:       { type: string, minLength: 1 }
        description: { type: string }
        status:      { type: string, enum: [todo, in_progress, blocked, review, done] }
        priority:    { type: string, enum: [low, medium, high, critical] }
        category:    { type: string, enum: [construction, electrical, av, it, network, acoustics, furnishing, signage, general] }
        dueDate:     { type: string }
    TaskUpdate:
      type: object
      properties:
        milestoneId: { type: integer }
        assigneeId:  { type: integer }
        title:       { type: string }
        description: { type: string }
        status:      { type: string, enum: [todo, in_progress, blocked, review, done] }
        priority:    { type: string, enum: [low, medium, high, critical] }
        category:    { type: string, enum: [construction, electrical, av, it, network, acoustics, furnishing, signage, general] }
        dueDate:     { type: string }
        completedAt: { type: string }
        note:        { type: string, description: "Optional note recorded with the change. When status changes, this becomes a status-change note." }
        noteAuthorName: { type: string }
...